Move Docker specific function out of the Kubelet
